+++ js/options.js
@@ -17,6 +17,7 @@
     document.getElementById('styleSelectedSimple').checked = extOptions.styleSelectedSimple;
     document.getElementById('styleSelectedFancy').checked = extOptions.styleSelectedFancy;
     document.getElementById('activateSearch').checked = extOptions.activateSearch;
+    document.getElementById('autoselectFirst').checked = extOptions.autoselectFirst;
     document.getElementById('selectTextInSearchbox').checked = extOptions.selectTextInSearchbox;
   }
 
@@ -27,6 +28,7 @@
     extOptions.styleSelectedSimple = document.getElementById('styleSelectedSimple').checked === true;
     extOptions.styleSelectedFancy = document.getElementById('styleSelectedFancy').checked === true;
     extOptions.activateSearch = document.getElementById('activateSearch').checked === true;
+    extOptions.autoselectFirst = document.getElementById('autoselectFirst').checked === true;
     extOptions.selectTextInSearchbox = document.getElementById('selectTextInSearchbox').checked === true;
     persistOptions();
   }